+++ source/renderer/app/stores/ada/AdaRedemptionStore.js
@@ -30,6 +30,7 @@ export default class AdaRedemptionStore extends Store {
   @observable email: ?string = null;
   @observable adaPasscode: ?string = null;
   @observable adaAmount: ?string = null;
+  @observable decryptionKey: ?string = null;
   @observable redemptionCode: string = '';
   @observable walletId: ?string = null;
   @observable error: ?LocalizableError = null;
@@ -49,6 +50,7 @@ export default class AdaRedemptionStore extends Store {
     actions.setEmail.listen(this._setEmail);
     actions.setAdaPasscode.listen(this._setAdaPasscode);
     actions.setAdaAmount.listen(this._setAdaAmount);
+    actions.setDecryptionKey.listen(this._setDecryptionKey);
     actions.redeemAda.listen(this._redeemAda);
     actions.redeemPaperVendedAda.listen(this._redeemPaperVendedAda);
     actions.adaSuccessfullyRedeemed.listen(this._onAdaSuccessfullyRedeemed);
@@ -96,9 +98,10 @@ export default class AdaRedemptionStore extends Store {
   _setCertificate = action(({ certificate }) => {
     this.certificate = certificate;
     this.isCertificateEncrypted = certificate.type !== 'application/pdf';
-    if (this.isCertificateEncrypted && !this.passPhrase) {
+    if (this.isCertificateEncrypted && (!this.passPhrase || !this.decryptionKey)) {
       this.redemptionCode = '';
       this.passPhrase = null;
+      this.decryptionKey = null;
       return; // We cannot decrypt it yet!
     }
     this._parseCodeFromCertificate();
@@ -128,8 +131,13 @@ export default class AdaRedemptionStore extends Store {
     this._parseCodeFromCertificate();
   });
 
+  _setDecryptionKey = action(({ decryptionKey } : { decryptionKey: string }) => {
+    this.decryptionKey = decryptionKey;
+    this._parseCodeFromCertificate();
+  });
+
   _parseCodeFromCertificate() {
-    if (this.redemptionType === 'regular') {
+    if (this.redemptionType === 'regular' || this.redemptionType === 'recoveryRegular') {
       if (!this.passPhrase && this.isCertificateEncrypted) return;
     }
     if (this.redemptionType === 'forceVended') {
@@ -137,16 +145,25 @@ export default class AdaRedemptionStore extends Store {
         return;
       }
     }
+    if (this.redemptionType === 'recoveryForceVended') {
+      if (!this.decryptionKey && this.isCertificateEncrypted) return;
+    }
     if (this.redemptionType === 'paperVended') return;
     if (this.certificate == null) throw new Error('Certificate File is required for parsing.');
     const path = this.certificate.path; // eslint-disable-line
     Logger.debug('Parsing ADA Redemption code from certificate: ' + path);
     let decryptionKey = null;
-    if (this.redemptionType === 'regular' && this.isCertificateEncrypted) {
+    if (
+      (this.redemptionType === 'regular' || this.redemptionType === 'recoveryRegular') &&
+      this.isCertificateEncrypted
+    ) {
       decryptionKey = this.passPhrase;
     }
     if (this.redemptionType === 'forceVended' && this.isCertificateEncrypted) {
       decryptionKey = [this.email, this.adaPasscode, this.adaAmount];
+    }
+    if (this.redemptionType === 'recoveryForceVended' && this.isCertificateEncrypted) {
+      decryptionKey = this.decryptionKey;
     }
     ipcRenderer.send(PARSE_REDEMPTION_CODE.REQUEST, path, decryptionKey, this.redemptionType);
   }
@@ -169,6 +186,7 @@ export default class AdaRedemptionStore extends Store {
     }
     this.redemptionCode = '';
     this.passPhrase = null;
+    this.decryptionKey = null;
   });
 
   _redeemAda = async ({ walletId, walletPassword } : {
@@ -231,6 +249,7 @@ export default class AdaRedemptionStore extends Store {
     this.showAdaRedemptionSuccessMessage = true;
     this.redemptionCode = '';
     this.passPhrase = null;
+    this.decryptionKey = null;
   });
 
   _onCloseAdaRedemptionSuccessOverlay = action(() => {
@@ -251,6 +270,7 @@ export default class AdaRedemptionStore extends Store {
     this.email = null;
     this.adaPasscode = null;
     this.adaAmount = null;
+    this.decryptionKey = null;
   });
 
   @action _reset = () => {
@@ -265,6 +285,7 @@ export default class AdaRedemptionStore extends Store {
     this.email = null;
     this.adaPasscode = null;
     this.adaAmount = null;
+    this.decryptionKey = null;
   };
 
 }