+++ source/renderer/app/types/redemptionTypes.js
@@ -1,2 +1,3 @@
 // @flow
-export type RedemptionTypeChoices = 'regular' | 'forceVended' | 'paperVended';
+export type RedemptionTypeChoices =
+  'regular' | 'forceVended' | 'paperVended' | 'recoveryRegular' | 'recoveryForceVended';